Modify threshold behaviour in deliverability source control
Previously, if the well pressure dropped below the threshold, the
deliverability control would activate, but it would not deactivate
again until the well pressure rose back above the threshold, even if
the computed flow rate was lower (i.e. more negative) than the
specified flow rate. This can happen if the specified flow rates are
time-dependent. Now deliverability is deactivated if the computed flow
rate is lower than the specified one.