Use custom SNES linesearch in timestepper
Standard PETSc basic linesearch does not check linesearch reason after
post-check routine, so setting it in there has no effect. Also, the
SNES function domain error is reset in the final function evaluation
of the linesearch, so setting domain error in the post-check has no
effect either. Here we set the linesearch reason in post-check and
check it afterwards - if it failed, set function domain error and do
not do the final function evaluation.